+++ pymotors/__init__.py
@@ -1,3 +1,4 @@
 from .stepper_base import StepperBase
 from .tic_stepper import TicStepper
 from .dc_base import DcBase, LimitedDc
+from .dc_array import DcArray