+++ pymotors/stepper_base.py
@@ -48,6 +48,12 @@ class StepperBase():
         self.rpm = rpm
         self.enable = self._enable_states['DISABLED']
 
+    def __del__(self):
+        if self.isMoving():
+            self.stop()
+        if self.enable:
+            self.enable = self._enable_states['DISABLED']
+
     @property
     def microsteps(self) -> float:
         """Fraction of a stepper motor full steps per pulse."""
@@ -57,7 +63,7 @@ class StepperBase():
     def microsteps(self, step_ratio: float):
         micros_per_full_step = 1 / step_ratio
         if self._checkMicrostep(micros_per_full_step):
-            old_micros = self.microsteps
+            old_micros = 1 / self.microsteps
             new_to_old = micros_per_full_step / old_micros
             self.rpm = self.rpm * new_to_old  # reset RPM for new micros
             self._setMicrostep(micros_per_full_step)
@@ -225,9 +231,9 @@ class StepperBase():
     def _typeSorter(self, val_type: str) -> int:
         """Convert string to int."""
         ret = self._unit_type['UNKNOWN']
-        if val_type in ('steps', 'Steps'):
+        if val_type in ('steps', 'Steps', 'STEPS'):
             ret = self._unit_type['STEPS']
-        elif val_type in ('dist', 'Dist'):
+        elif val_type in ('dist', 'Dist', 'DIST'):
             ret = self._unit_type['DIST']
         else:
             warnings.warn("Expected `dist` or `steps`.")