+++ CHANGES.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. This module adheres to [Semantic Versioning](https://semver.org/).
 
+## 1.3.0
+- Remove timers from DcBase and LimitedDc
+- Add command to set limits in LimitedDc
+- Make LimitedDc checkLimits method public now that timers no longer check state in background
+- Update tests for DcBase and LimitedDc
+- Remove Raspberry Pi dependencies in DC class
+- Add logging to DC classes
+
+## 1.2.0
+
+- Add zeroing command in TicStepper to set current position to 0
+- Override isMoving in TicStepper to check current velocity
+
+## 1.1.0
+- Close serial and I2C ports upon deleting the Tic.com object
+- Stop and disable StepperBase during deletion
+
+## 1.0.2
+- Patch StepperBase microsteps method to correctly handle successive calls
+
 ## 1.0.1
 - Minor corrections to StepperBase and TicStepper documentation.